pyproject.toml
[tool.poetry]
name = "careergpt"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12.2"
django = "4.2"
python-dotenv = "^1.0.1"
openai = "^1.23.2"
pymupdf = "^1.24.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
